refactor(e2e): isolate fault test modules
